# Speed up ingestion using bulk insert / updates