… sv_mini_atlas/tests/test_importer.py