@jacobwegner 2020-02-19
Deployed to  https://mini-stack-a-spike-spee-wkno11.herokuapp.com/graphql/