@jacobwegner 2020-02-27
![](https://github.trello.services/images/mini-trello-icon.png) [Improve ingestion speed](https://trello.com/c/llhOXBlH/38-improve-ingestion-speed)