@jacobwegner 2020-03-10
@jhrr thanks for your work here; I want to add some docstrings but then will flag for a final review.

I just pushed a small change set where I removed some dead code and added simple assertions for `numchild` counts.